tinyvlm/requirements.txt
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
pillow>=10.0.0
tensorboard>=2.15.0
safetensors>=0.4.0
pyyaml>=6.0
tqdm>=4.66.0
